+++ core/src/main/kotlin/io/github/setheclark/intellij/system/process/ProcessExecutor.kt
@@ -1,4 +1,4 @@
-package io.github.setheclark.intellij.process
+package io.github.setheclark.intellij.system.process
 
 /**
  * Interface for executing system processes.