+++ plugin/src/main/kotlin/io/github/setheclark/intellij/di/AppGraph.kt
@@ -8,11 +8,13 @@ import dev.zacsweers.metro.SingleIn
 import io.github.setheclark.intellij.flocon.FloconBindingContainer
 import io.github.setheclark.intellij.network.InMemoryNetworkDataSource
 import io.github.setheclark.intellij.network.NetworkDataSource
-import io.github.setheclark.intellij.process.ProcessExecutor
-import io.github.setheclark.intellij.process.SystemProcessExecutor
+import io.github.setheclark.intellij.system.process.ProcessExecutor
+import io.github.setheclark.intellij.system.process.SystemProcessExecutor
 import io.github.setheclark.intellij.services.ApplicationServiceDelegate
 import io.github.setheclark.intellij.settings.NetworkStorageSettingsProvider
 import io.github.setheclark.intellij.settings.NetworkStorageSettingsProviderImpl
+import io.github.setheclark.intellij.system.Environment
+import io.github.setheclark.intellij.system.IntellijEnvironment
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.serialization.json.Json
 
@@ -34,6 +36,9 @@ interface AppGraph : ProjectGraph.Factory {
 
     @Binds
     val NetworkStorageSettingsProviderImpl.bind: NetworkStorageSettingsProvider
+
+    @Binds
+    val IntellijEnvironment.bind: Environment
 
     @Provides
     @SingleIn(AppScope::class)