+++ plugin/src/main/kotlin/io/github/setheclark/intellij/system/IntellijEnvironment.kt
@@ -1,0 +1,14 @@
+package io.github.setheclark.intellij.system
+
+import com.intellij.util.EnvironmentUtil
+import dev.zacsweers.metro.AppScope
+import dev.zacsweers.metro.Inject
+import dev.zacsweers.metro.SingleIn
+
+@Inject
+@SingleIn(AppScope::class)
+class IntellijEnvironment : Environment {
+    override fun getEnvironmentVariable(key: String): String? {
+        return EnvironmentUtil.getValue(key)
+    }
+}