+++ plugin/src/main/kotlin/io/github/setheclark/intellij/system/process/SystemProcessExecutor.kt
@@ -1,4 +1,4 @@
-package io.github.setheclark.intellij.process
+package io.github.setheclark.intellij.system.process
 
 import co.touchlab.kermit.Logger
 import com.intellij.execution.configurations.GeneralCommandLine